Attempting forced merge

+++ source/calculators/onegapignore.h
@@ -32,13 +32,13 @@ public:
 		
 		string seqA = A.getAligned();
 		string seqB = B.getAligned();
-		int alignLength = seqA.length();
+		int alignLength = (int)seqA.length();
 
 		// this assumes that sequences start and end with '.'s instead of'-'s.
 		for(int i=0;i<alignLength;i++){
 			if(seqA[i] != '.' && seqB[i] != '.' && seqA[i] != '-' && seqB[i] != '-' ){
 				start = i;
-//				cout << "start: " << start << endl;
+
 				overlap = true;
 				break;
 			}
@@ -46,7 +46,7 @@ public:
 		for(int i=alignLength-1;i>=0;i--){
 			if(seqA[i] != '.' && seqB[i] != '.' && seqA[i] != '-' && seqB[i] != '-' ){
 				end = i;
-//				cout << "end: " << end << endl;
+
 				overlap = true;
 				break;
 			}
